Move the `env-client` module under the `@bayou` namespace.
