+++ client/js/index.js
@@ -10,7 +10,7 @@
  * system.
  */
 
-import { ClientEnv } from 'env-client';
+import { ClientEnv } from '@bayou/env-client';
 import { Logger } from '@bayou/see-all';
 import { ClientSink } from '@bayou/see-all-client';
