+++ client/js/tests/index.js
@@ -6,7 +6,7 @@
  * Top-level entry point for client tests.
  */
 
-import { ClientEnv } from 'env-client';
+import { ClientEnv } from '@bayou/env-client';
 import { Logger } from '@bayou/see-all';
 import { ClientSink } from '@bayou/see-all-client';
 import { Tests } from '@bayou/testing-client';
