+++ client/package.json
@@ -19,7 +19,7 @@
     "@bayou/api-common": "local",
     "@bayou/app-common": "local",
     "doc-client": "local",
-    "env-client": "local",
+    "@bayou/env-client": "local",
     "@bayou/see-all-client": "local",
     "@bayou/testing-client": "local"
   }
