+++ local-modules/@bayou/env-client/package.json
@@ -1,9 +1,8 @@
 {
-  "name": "env-client",
+  "name": "@bayou/env-client",
   "version": "1.0.0",
 
   "dependencies": {
-    "@bayou/see-all": "local",
     "@bayou/typecheck": "local",
     "@bayou/util-common": "local"
   }
